requirements.txt.txt
streamlit
pandas
openpyxl
python-docx
python-pptx
google-generativeai
streamlit-mic-recorder
requests
reportlab
pdfplumber
moviepy